requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
faster-whisper>=1.1.0
numpy>=1.26.0
httpx>=0.27.0
openai>=1.50.0
python-dotenv>=1.0.0
